separate RobotGripperSystem from RobotArmSystem
